third_party/blink/web_tests/svg/as-image/svg-container-size-after-reload-expected.txt

Test for WK112651: SVG image container size should not survive reloads.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS image2.width is 100
PASS successfullyParsed is true

TEST COMPLETE
